in preparazione per MLFQ

+++ codice/statistics/stat.h
@@ -6,8 +6,8 @@
 typedef struct {
 	int num_proc; //il processo con pid i corrisponde a (num_proc - 1) 
 	int num_core;
-	float* waiting_time; //per ogni processo sarà assegnato un tempo totale di attesa prima di essere stato schedulato
-	float* completing_time; //per ogni processo sarà assegnato un tempo totale di completamento
+	int* waiting_time; //per ogni processo sarà assegnato un tempo totale di attesa prima di essere stato schedulato
+	int* completing_time; //per ogni processo sarà assegnato un tempo totale di completamento
 	float medium_wait_time; //tempo medio di attesa
 	float medium_complete_time; //tempo medio di completamento
 } StatisticsType;
